% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centile.R
\name{centile}
\alias{centile}
\title{Calculate the percentile ratios}
\usage{
centile(data_vector, top_perc, bot_perc)
}
\arguments{
\item{data_vector}{A numeric vector.}

\item{top_perc}{The top percentile you will put in your equation.}

\item{bot_perc}{The bot perncetile you will put in your equation.}
}
\value{
Result of diving top_perc into bot_perc.
}
\description{
Takes in a vector and two percentile then divides them.
}
