% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centile.R
\name{mcloone}
\alias{mcloone}
\title{Calculates McLoone index.}
\usage{
mcloone(data_vector)
}
\arguments{
\item{data_vector}{A numeric vector.}
}
\value{
How values differ from median.
}
\description{
Calculates McLoone index.
}
