% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centile.R
\name{palma}
\alias{palma}
\title{Calculates the palma ratio.}
\usage{
palma(data_vector)
}
\arguments{
\item{data_vector}{A numeric vector.}
}
\value{
Result of first 10 percentile / last 40 percentile.
}
\description{
Calculates the palma ratio.
}
