% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centile.R
\name{relvar}
\alias{relvar}
\title{Calculates the relative variance.}
\usage{
relvar(data_vector)
}
\arguments{
\item{data_vector}{A numeric vector.}
}
\value{
Relative variance.
}
\description{
Calculates the relative variance.
}
